progress on services
